test: Fixture 구현하여 테스트 리팩토링
