+++ src/test/java/roomescape/support/dto/TokenCookieDto.java
@@ -1,0 +1,4 @@
+package roomescape.support.dto;
+
+public record TokenCookieDto(String accessTokenCookie, String refreshTokenCookie) {
+}
