--- src/test/java/roomescape/support/model/TokenCookieDto.java
@@ -1,4 +1,0 @@
-package roomescape.support.model;
-
-public record TokenCookieDto(String accessTokenCookie, String refreshTokenCookie) {
-}
